# Remove sentinels as they race with multi threading