--- a/src/engine/Engine.h
+++ b/src/engine/Engine.h
@@ -125,31 +125,6 @@ class Engine {
       return;
     }
 
-    // Cast away constness so we can add sentinels that will be removed
-    // in the end and create and add those sentinels.
-    IdTable& l1 = const_cast<IdTable&>(dynV);
-    IdTable& l2 = const_cast<IdTable&>(dynFilter);
-
-    Id sent1 = std::numeric_limits<Id>::max();
-    Id sent2 = std::numeric_limits<Id>::max() - 1;
-    Id sentMatch = std::numeric_limits<Id>::max() - 2;
-    auto elem1 = l1[0];
-    auto elem2 = l2[0];
-    auto match1 = l1[0];
-    auto match2 = l2[0];
-
-    l1.emplace_back();
-    l1.back()[fc1] = sentMatch;
-    l1.back()[fc2] = sentMatch;
-    l1.emplace_back();
-    l1.back()[fc1] = sent1;
-
-    l2.emplace_back();
-    l2.back()[0] = sentMatch;
-    l2.back()[1] = sentMatch;
-    l2.emplace_back();
-    l2.back()[fc2] = sent2;
-
     const IdTableStatic<IN_WIDTH> v = dynV.asStaticView<IN_WIDTH>();
     const IdTableStatic<FILTER_WIDTH> filter =
         dynFilter.asStaticView<FILTER_WIDTH>();
@@ -159,12 +134,18 @@ class Engine {
     size_t i = 0;
     size_t j = 0;
 
-    while (v(i, fc1) < sent1) {
+    while (i < v.size() && j < filter.size()) {
       while (v(i, fc1) < filter(j, 0)) {
         ++i;
+        if (i >= v.size()) {
+          goto finish;
+        }
       }
       while (filter(j, 0) < v(i, fc1)) {
         ++j;
+        if (j >= filter.size()) {
+          goto finish;
+        }
       }
       while (v(i, fc1) == filter(j, 0)) {
         // fc1 match, create cross-product
@@ -172,21 +153,23 @@ class Engine {
         if (v(i, fc2) == filter(j, 1)) {
           result.push_back(v, i);
           ++i;
-          if (i == v.size()) break;
+          if (i >= v.size()) {
+            break;
+          }
         } else if (v(i, fc2) < filter(j, 1)) {
           ++i;
-          if (i == v.size()) break;
+          if (i >= v.size()) {
+            break;
+          }
         } else {
           ++j;
-          if (j == filter.size()) break;
+          if (j >= filter.size()) {
+            break;
+          }
         }
       }
     }
-
-    // Remove sentinels
-    l1.resize(l1.size() - 2);
-    l2.resize(l2.size() - 2);
-    result.pop_back();
+  finish:
     *dynResult = result.moveToDynamic();
 
     LOG(DEBUG) << "Filter done, size now: " << dynResult->size()
@@ -283,28 +266,10 @@ class Engine {
       return;
     }
 
-    // Cast away constness so we can add sentinels that will be removed
-    // in the end and create and add those sentinels.
-    IdTable& l1 = const_cast<IdTable&>(dynA);
-    IdTable& l2 = const_cast<IdTable&>(dynB);
-
-    Id sent1 = std::numeric_limits<Id>::max();
-    Id sent2 = std::numeric_limits<Id>::max() - 1;
-    Id sentMatch = std::numeric_limits<Id>::max() - 2;
-
-    l1.push_back();
-    l1.back()[jc1] = sentMatch;
-    l1.push_back();
-    l1.back()[jc1] = sent1;
-
-    l2.push_back();
-    l2.back()[jc2] = sentMatch;
-    l2.push_back();
-    l2.back()[jc2] = sent2;
-
     // Update the views
     const IdTableStatic<L_WIDTH> a = dynA.asStaticView<L_WIDTH>();
     const IdTableStatic<R_WIDTH> b = dynB.asStaticView<R_WIDTH>();
+    // TODO(schnelle) get rid of goto by using return
 
     // Cannot just switch l1 and l2 around because the order of
     // items in the result tuples is important.
@@ -316,13 +281,22 @@ class Engine {
       // Intersect both lists.
       size_t i = 0;
       size_t j = 0;
-      while (a(i, jc1) < sent1) {
+      // while (a(i, jc1) < sent1) {
+      while (i < a.size() && j < b.size()) {
         while (a(i, jc1) < b(j, jc2)) {
           ++i;
+          if (i >= a.size()) {
+            goto finish;
+          }
         }
+
         while (b(j, jc2) < a(i, jc1)) {
           ++j;
+          if (j >= b.size()) {
+            goto finish;
+          }
         }
+
         while (a(i, jc1) == b(j, jc2)) {
           // In case of match, create cross-product
           // Always fix a and go through b.
@@ -334,6 +308,8 @@ class Engine {
             for (size_t h = 0; h < aCols; h++) {
               result(backIndex, h) = a(i, h);
             }
+            // TODO(schnelle): Maybe we can tighten the loop
+            // below by doing smaller/larger separately
             for (size_t h = 0; h < b.cols(); h++) {
               if (h < jc2) {
                 result(backIndex, h + aCols) = b(j, h);
@@ -342,8 +318,14 @@ class Engine {
               }
             }
             ++j;
+            if (j >= b.size()) {
+              break;
+            }
           }
           ++i;
+          if (i >= a.size()) {
+            goto finish;
+          }
           // If the next i is still the same, reset j.
           if (a(i, jc1) == b(keepJ, jc2)) {
             j = keepJ;
@@ -351,10 +333,7 @@ class Engine {
         }
       }
     }
-    // Remove sentinels
-    l1.resize(l1.size() - 2);
-    l2.resize(l2.size() - 2);
-    result.pop_back();
+  finish:
     *dynRes = result.moveToDynamic();
 
     LOG(DEBUG) << "Join done.\n";
@@ -429,23 +408,30 @@ class Engine {
                                           const IdTableStatic<R_WIDTH>& l2,
                                           size_t jc2,
                                           IdTableStatic<OUT_WIDTH>* result) {
+    // TODO(schnelle) get rid of goto by using return
     LOG(DEBUG) << "Galloping case.\n";
     size_t i = 0;
     size_t j = 0;
-    Id sent1 = std::numeric_limits<Id>::max();
-    while (l1(i, jc1) < sent1) {
+    // TODO(schnelle) when we have fixed cols we could constexpr if this
+    // and use stack allocation which should be much faster
+    Id* val = new Id[l2.cols()];
+    while (i < l1.size() && j < l2.size()) {
       while (l1(i, jc1) < l2(j, jc2)) {
         ++i;
+        if (i >= l1.size()) {
+          goto finish;
+        }
       }
       if (l2(j, jc2) < l1(i, jc1)) {
-        Id* val = new Id[l2.cols()];
         val[jc2] = l1(i, jc1);
         j = std::lower_bound(l2.begin() + j, l2.end(), val,
                              [jc2](const auto& l, const auto& r) -> bool {
                                return l[jc2] < r[jc2];
                              }) -
             l2.begin();
-        delete[] val;
+        if (j >= l2.size()) {
+          goto finish;
+        }
       }
       while (l1(i, jc1) == l2(j, jc2)) {
         // In case of match, create cross-product
@@ -466,14 +452,22 @@ class Engine {
             }
           }
           ++j;
+          if (j >= l2.size()) {
+            goto finish;
+          }
         }
         ++i;
+        if (i >= l1.size()) {
+          goto finish;
+        }
         // If the next i is still the same, reset j.
         if (l1(i, jc1) == l2(keepJ, jc2)) {
           j = keepJ;
         }
       }
     }
+  finish:
+    delete[] val;
   };
 
   template <int L_WIDTH, int R_WIDTH, int OUT_WIDTH>
@@ -482,23 +476,30 @@ class Engine {
                                          const IdTableStatic<R_WIDTH>& l2,
                                          size_t jc2,
                                          IdTableStatic<OUT_WIDTH>* result) {
+    // TODO(schnelle) get rid of goto by using return
     LOG(DEBUG) << "Galloping case.\n";
     size_t i = 0;
     size_t j = 0;
-    Id sent1 = std::numeric_limits<Id>::max();
-    while (l1(i, jc1) < sent1) {
+    // TODO(schnelle) when we have fixed cols we could constexpr if this
+    // and use stack allocation which should be much faster
+    Id* val = new Id[l1.cols()];
+    while (i < l1.size() && j < l2.size()) {
       if (l2(j, jc2) > l1(i, jc1)) {
-        Id* val = new Id[l1.cols()];
         val[jc1] = l2(j, jc2);
         i = std::lower_bound(l1.begin() + i, l1.end(), val,
                              [jc1](const auto& l, const auto& r) -> bool {
                                return l[jc1] < r[jc1];
                              }) -
             l1.begin();
-        delete[] val;
+        if (i >= l1.size()) {
+          goto finish;
+        }
       }
       while (l1(i, jc1) > l2(j, jc2)) {
         ++j;
+        if (j >= l2.size()) {
+          goto finish;
+        }
       }
       while (l1(i, jc1) == l2(j, jc2)) {
         // In case of match, create cross-product
@@ -519,13 +520,21 @@ class Engine {
             }
           }
           ++j;
+          if (j >= l2.size()) {
+            goto finish;
+          }
         }
         ++i;
+        if (i >= l1.size()) {
+          goto finish;
+        }
         // If the next i is still the same, reset j.
         if (l1(i, jc1) == l2(keepJ, jc2)) {
           j = keepJ;
         }
       }
     }
+  finish:
+    delete[] val;
   };
 };