--- a/src/engine/Engine.h
+++ b/src/engine/Engine.h
@@ -231,7 +231,7 @@ class Engine {
 
             ++j;
             if (j >= b.size()) {
-              break;
+              goto finish;
             }
           }
           ++i;