--- a/src/engine/Server.h
+++ b/src/engine/Server.h
@@ -32,6 +32,8 @@ class Server {
         _engine(),
         _initialized(false) {}
 
+  virtual ~Server();
+
   typedef ad_utility::HashMap<string, string> ParamValueMap;
 
   // Initialize the server.