… src/engine/Engine.cpp
… src/engine/MultiColumnJoin.cpp
… src/index/FTSAlgorithms.cpp
… src/index/Index.cpp
… src/index/MetaDataIterator.h
… src/util/Simple8bCode.h
… src/util/Socket.h