Also add some TODOs for ideas how to get joins faster and we may still remove the `goto`s