@niklas88 2019-03-28
> Idea: We could handle the sentinels inside the IdTable
> by using a `finalizeSentinels` class, that writes a maximums beyond bounds. Afterwards we have the sentinel in the off-by-one area and can use them with the old code.

Let's first see how much performance the removal of sentinels costs us. I think while the code looks cleaner they are also harder to reason about so they really need to pay for themselves with performance