@niklas88 2019-03-28
@joka921 ok this is interesting. Turns out, this branch is actually 22% faster than current master and still 18% faster than the master branch after applying your change to the use of `std::lower_bound`
With sentinels and new/delete:

    Average completion moves: 4.52 (vs 8.51 manually)
    Executed 272 queries in 15032.706 seconds
    with the following timings

    Average: 55.27 ms/query
    Max: 837.45 ms/query
    90th percentile: 106.33 ms/query
    70th percentile: 71.68 ms/query
    50th percentile: 8.52 ms/query

With sentinels but no new/delete:

    Average completion moves: 4.52 (vs 8.51 manually)
    Executed 272 queries in 14579.671 seconds
    with the following timings

    Average: 53.60 ms/query
    Max: 833.57 ms/query
    90th percentile: 109.39 ms/query
    70th percentile: 70.67 ms/query
    50th percentile: 7.94 ms/query

Without sentinels:

    Average completion moves: 4.52 (vs 8.51 manually)
    Executed 287 queries in 13004.729 seconds
    with the following timings

    Average: 45.31 ms/query
    Max: 710.91 ms/query
    90th percentile: 93.13 ms/query
    70th percentile: 27.60 ms/query
    50th percentile: 5.30 ms/query

So it looks like sentinels are actually bad for performance. I guess without them the compiler just knows more and the branch predictor has no problems guessing the branches that one safes anyway.