@niklas88 2019-03-28
216 fewer lines of code to break with all the same features but ~22% faster. Inflationary…