@niklas88 2019-03-28
I've found another broken use of sentinels in the Text stuff. Will have to fix that tomorrow as the straightforward versions doesn't seem to work yet